move types to folder
